refactor: 해당 인라인 callback 함수를 따로 분리하여 명시적으로 정의
